fix: ignore Expect header when signing requests
This would resolve a common issue with s3 proxy configuration. 

In some s3-compliant storage providers, there are multiple proxy hops between the client and server-side signature verification. Often, these are ingress gateways like Istio (Envoy), whose default behavior as of 1.21 is to handle 100-continue headers as if it's the final destination, stripping them from the request after sending the 100-continue to the client. 

The issue arises when the first proxy hop, which handles the 100-continue logic, is not the same server that verifies the AWS Signature. 

In cases like this, the Authorization header will include `expect` in the list of signed headers, but the request that the signature verifier sees does not include that header.

This leads to signature verification errors when the server attempts to set up the canonical request with `Expect: ` instead of `Expect: 100-continue`. 

This PR adds Expect to the list of ignored headers when signing, allowing requests to be handled regardless of intervening proxy hops.